\name{SA23204167-package}
\alias{SA23204167-package}
\docType{package}
\title{
  Implement parts of GFM-TW algorithm and related simulations
}
\description{
  An R package that implements part of the simulation section of a nonlinear missing data completion algorithm called GFM-TW.
}
\author{
Xinyu Li
Maintainer: Xinyu Li <lx20160727@mail.ustc.edu.cn>
}
\keyword{ package }
